move useQuery to use-plugins
